bca-market-services/src/requirements.txt
mypy
thriftpy2
psycopg2
types-psycopg2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

